[WASI-NN] Apply the graph and context management mechanism for all backends.

+++ plugins/wasi_nn/wasinn_mlx.h
@@ -37,15 +37,15 @@ struct Graph {
   BasePrompt Prmopt;
 };
 struct Context {
-  Context(size_t Gid, Graph &) noexcept : GraphId(Gid) {}
-  size_t GraphId;
+  Context(uint32_t Gid, Graph &) noexcept : GraphId(Gid) {}
+  uint32_t GraphId;
   std::string Inputs;
   std::string Outputs;
 };
 #else
 struct Graph {};
 struct Context {
-  Context(size_t, Graph &) noexcept {}
+  Context(uint32_t, Graph &) noexcept {}
 };
 #endif
